assessment-service/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
google-generativeai
google-cloud-storage
pydub
python-multipart